requirements.txt
pillow
replicate
gradio-imageslider==0.0.13
gradio
numpy
torch
opencv-python
torchvision
streamlit